+++ backend/src/main/java/com/example/backend/application/usecases/reports/ReportGenerationUseCase.java
@@ -1,12 +1,14 @@
 package com.example.backend.application.usecases.reports;
 
-import com.example.backend.application.dto.*;
+import com.example.backend.application.dto.reports.DailyReportDto;
+import com.example.backend.presentation.controllers.reports.DailyReportCreateRequestDto;
+import com.example.backend.presentation.controllers.reports.DailyReportUpdateRequestDto;
 import com.example.backend.presentation.dto.reports.ReportGenerationRequestDto;
 import com.example.backend.presentation.dto.reports.ReportGenerationResponseDto;
 import com.example.backend.presentation.dto.reports.ReportRegenerationRequestDto;
-import com.example.backend.application.usecases.GitHubUseCase;
-import com.example.backend.application.usecases.NotionUseCase;
-import com.example.backend.application.usecases.ToggleTrackUseCase;
+import com.example.backend.application.usecases.external.github.GitHubUseCase;
+import com.example.backend.application.usecases.external.notion.NotionUseCase;
+import com.example.backend.application.usecases.external.toggle.ToggleTrackUseCase;
 import com.example.backend.domain.reports.IReportGenerationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
