--- backend/src/main/java/com/example/backend/common/jooq/CustomGeneratorStrategy.java
@@ -1,12 +1,0 @@
-package com.example.backend.common.jooq;
-
-import org.jooq.codegen.DefaultGeneratorStrategy;
-import org.jooq.meta.Definition;
-
-public class CustomGeneratorStrategy extends DefaultGeneratorStrategy {
-  @Override
-  public String getJavaClassName(Definition definition, Mode mode) {
-    String defaultName = super.getJavaClassName(definition, mode);
-    return "J" + defaultName;
-  }
-}
