+++ backend/src/test/java/com/example/backend/application/usecases/ReportGenerationUseCaseTest.java
@@ -1,6 +1,10 @@
 package com.example.backend.application.usecases;
 
-import com.example.backend.application.dto.*;
+import com.example.backend.application.dto.reports.DailyReportDto;
+import com.example.backend.application.usecases.external.github.GitHubUseCase;
+import com.example.backend.application.usecases.external.notion.NotionUseCase;
+import com.example.backend.application.usecases.external.toggle.ToggleTrackUseCase;
+import com.example.backend.presentation.controllers.reports.DailyReportCreateRequestDto;
 import com.example.backend.presentation.dto.reports.ReportGenerationRequestDto;
 import com.example.backend.presentation.dto.reports.ReportGenerationResponseDto;
 import com.example.backend.application.usecases.reports.ReportGenerationUseCase;
